- 16 bit for baroTemperature instead of 32 bit
git-svn-id: http://multiwii.googlecode.com/svn/trunk/MultiWii_shared@1624 02679b44-d973-9852-f2fa-63770883b36c

+++ IMU.cpp
@@ -327,7 +327,7 @@ uint8_t getEstimatedAltitude(){
 
   if(calibratingB > 0) {
     logBaroGroundPressureSum = log(baroPressureSum);
-    baroGroundTemperatureScale = (baroTemperature + 27315) *  29.271267f;
+    baroGroundTemperatureScale = ((int32_t)baroTemperature + 27315) *  29.271267f;
     calibratingB--;
   }
