+++ MultiWii.h
@@ -70,7 +70,7 @@ extern int16_t angle[2];
 
 #if BARO
   extern int32_t baroPressure;
-  extern int32_t baroTemperature;
+  extern int16_t baroTemperature; // temp in 0.01 deg
   extern int32_t baroPressureSum;
 #endif
